IS 1182 fix tests
