+++ core/dkg/fair/utils.py
@@ -26,6 +26,7 @@ from core.dkg.fair.client import FairDKGClient
 from core.dkg.structures import DKGStep
 from core.dkg.utils import (
     DkgError,
+    DKGKeyGenerationError,
     BroadcastResult,
     BROADCAST_DATA_SEARCH_SLEEP,
     sync_broadcast_data,
@@ -146,3 +147,27 @@ def check_dkg_id_with_exception(dkg_client: FairDKGClient):
     if not dkg_client.check_round_id():
         logger.info('Restarting DKG: round id mismatch.')
         raise DkgError('Restarting DKG: round id mismatch.')
+
+
+def generate_bls_keys(dkg_client):
+    try:
+        if not dkg_client.is_bls_key_generated():
+            encrypted_bls_key = dkg_client.generate_bls_key()
+            logger.info(f'Node`s encrypted bls key is: {encrypted_bls_key}')
+        else:
+            logger.info('BLS key exists. Fetching')
+            dkg_client.fetch_bls_public_key()
+
+        bls_public_keys = dkg_client.get_bls_public_keys()
+        common_public_key = dkg_client.get_common_bls_public_key()
+    except Exception as err:
+        raise DKGKeyGenerationError(err)
+    dkg_client.last_completed_step = DKGStep.KEY_GENERATION
+    return {
+        'common_public_key': common_public_key,
+        'public_key': dkg_client.public_key,
+        'bls_public_keys': bls_public_keys,
+        't': dkg_client.t,
+        'n': dkg_client.n,
+        'key_share_name': dkg_client.bls_name,
+    }
